+++ packages/framer/src/checkout.tsx
@@ -1,7 +1,13 @@
 import {
+	type WhopCheckoutEmbedThemeOptions,
+	type WhopEmbeddedCheckoutPrefillOptions,
 	type WhopEmbeddedCheckoutStyleOptions,
 	WhopCheckoutEmbed as WhopReactCheckoutEmbed,
 } from "@whop/react/checkout";
+import {
+	type AccentColor,
+	accentColorValues,
+} from "@whop/react/checkout/colors";
 import { ControlType, type PropertyControls } from "framer";
 import React, { useMemo } from "react";
 import type { ReactNode } from "react";
@@ -17,6 +23,8 @@ export interface WhopFramerCheckoutEmbedProps {
 	containerPaddingTop?: number;
 	containerPaddingBottom?: number;
 	containerPaddingY?: number;
+	prefillEmail?: string;
+	accentColor?: AccentColor | "default";
 }
 
 export function WhopFramerCheckoutEmbed(props: WhopFramerCheckoutEmbedProps) {
@@ -33,6 +41,17 @@ export function WhopFramerCheckoutEmbed(props: WhopFramerCheckoutEmbedProps) {
 		props.containerPaddingTop,
 		props.containerPaddingY,
 	]);
+	const prefill: WhopEmbeddedCheckoutPrefillOptions = useMemo(() => {
+		return {
+			email: props.prefillEmail,
+		};
+	}, [props.prefillEmail]);
+	const themeOptions: WhopCheckoutEmbedThemeOptions = useMemo(() => {
+		return {
+			accentColor:
+				props.accentColor === "default" ? undefined : props.accentColor,
+		};
+	}, [props.accentColor]);
 	return (
 		<WhopReactCheckoutEmbed
 			planId={props.planId}
@@ -42,6 +61,8 @@ export function WhopFramerCheckoutEmbed(props: WhopFramerCheckoutEmbedProps) {
 			onComplete={props.onComplete}
 			fallback={props.fallback}
 			styles={styles}
+			prefill={prefill}
+			themeOptions={themeOptions}
 		/>
 	);
 }
@@ -103,5 +124,24 @@ export const propertyControls: PropertyControls<WhopFramerCheckoutEmbedProps> =
 		containerPaddingY: {
 			type: ControlType.Number,
 			description: "The vertical padding of the checkout embed container",
+		},
+		prefillEmail: {
+			type: ControlType.String,
+			description: "The email to prefill in the checkout embed",
+		},
+		accentColor: {
+			type: ControlType.Enum,
+			displaySegmentedControl: true,
+			defaultValue: "default",
+			segmentedControlDirection: "vertical",
+			options: ["default", ...accentColorValues],
+			optionTitles: [
+				"Default",
+				...accentColorValues.map((color) => {
+					const [firstChar, ...rest] = color;
+					return `${firstChar.toUpperCase()}${rest.join("")}`;
+				}),
+			],
+			description: "The accent color you want to use in the embed",
 		},
 	};