+++ packages/checkout/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @whop/checkout
 
+## 0.0.34
+
+### Patch Changes
+
+- 9a6c80c: add option to change checkout embed accent color
+
+## 0.0.33
+
+### Patch Changes
+
+- eda990d: add option to prefill email in checkout embed
+- cf63220: fix onComplete callback
+
 ## 0.0.32
 
 ### Patch Changes