+++ packages/framer/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @whop/framer
 
+## 0.0.4
+
+### Patch Changes
+
+- 9a6c80c: add option to change checkout embed accent color
+- Updated dependencies [9a6c80c]
+  - @whop/react@0.2.21
+
+## 0.0.3
+
+### Patch Changes
+
+- eda990d: add option to prefill email in checkout embed
+- Updated dependencies [eda990d]
+- Updated dependencies [f6b475f]
+  - @whop/react@0.2.20
+
 ## 0.0.2
 
 ### Patch Changes